+++ code/broker-server/src/main/java/com/google/cloud/broker/apps/brokerserver/validation/ProxyUserValidation.java
@@ -1,4 +1,4 @@
-// Copyright 2019 Google LLC
+// Copyright 2020 Google LLC
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -94,6 +94,7 @@ public class ProxyUserValidation {
 
     public static void validateImpersonator(String impersonator, String impersonated) {
         String mappedImpersonated = AbstractUserMapper.getInstance().map(impersonated);
+        Validation.validateEmail(mappedImpersonated);
         MDC.put("impersonated_user", impersonated);
         MDC.put("impersonated_user_mapped", mappedImpersonated);
         List<? extends Config> proxyConfigs = AppSettings.getInstance().getConfigList(AppSettings.PROXY_USERS);