+++ code/broker-server/src/main/java/com/google/cloud/broker/apps/brokerserver/validation/Validation.java
@@ -12,12 +12,30 @@
 package com.google.cloud.broker.apps.brokerserver.validation;
 
 import java.util.*;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import io.grpc.Status;
 
 import com.google.cloud.broker.settings.AppSettings;
 
 public class Validation {
+
+    public static void validateParameterIsEmpty(String parameter, String value) {
+        if (value.length() > 0) {
+            throw Status.INVALID_ARGUMENT
+                .withDescription(String.format("Request's parameter `%s` must be empty", parameter))
+                .asRuntimeException();
+        }
+    }
+
+    public static void validateParameterIsEmpty(String parameter, List<String> values) {
+        if (values.size() > 0) {
+            throw Status.INVALID_ARGUMENT
+                .withDescription(String.format("Request's parameter `%s` must be empty", parameter))
+                .asRuntimeException();
+        }
+    }
 
     public static void validateParameterNotEmpty(String parameter, String value) {
         if (value.length() == 0) {
@@ -43,6 +61,14 @@ public class Validation {
             throw Status.PERMISSION_DENIED
                 .withDescription(String.format("`[%s]` are not whitelisted scopes", String.join(",", scopes)))
                 .asRuntimeException();
+        }
+    }
+
+    public static void validateEmail(String email) {
+        Pattern parser = Pattern.compile("([a-zA-Z0-9\\.-]+)@([a-zA-Z0-9\\.-]+)");
+        Matcher match = parser.matcher(email);
+        if (!match.matches()) {
+            throw new IllegalArgumentException("Invalid email: " + email);
         }
     }