+++ code/broker-server/src/test/java/com/google/cloud/broker/apps/brokerserver/validation/ValidationTest.java
@@ -1,3 +1,14 @@
+// Copyright 2020 Google LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package com.google.cloud.broker.apps.brokerserver.validation;
 
 import java.util.List;
@@ -65,6 +76,19 @@ public class ValidationTest {
         } catch (StatusRuntimeException e) {
             assertEquals(Status.PERMISSION_DENIED.getCode(), e.getStatus().getCode());
             assertEquals("`[https://www.googleapis.com/auth/bigtable.data.readonly]` are not whitelisted scopes", e.getStatus().getDescription());
+        }
+    }
+
+    @Test
+    public void validateEmail() {
+        Validation.validateEmail("alice@example.com");
+        Validation.validateEmail("alice-shadow@my-project.iam.gserviceaccount.com");
+        for (String value : new String[]{"alice", "alice@", "@example.com", "xxx()@xxx"})
+        try {
+            Validation.validateEmail(value);
+            fail();
+        } catch (IllegalArgumentException e) {
+            // Expected
         }
     }