+++ code/core/src/main/java/com/google/cloud/broker/usermapping/MapUser.java
@@ -1,4 +1,4 @@
-// Copyright 2019 Google LLC
+// Copyright 2020 Google LLC
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -14,7 +14,8 @@ package com.google.cloud.broker.usermapping;
 public class MapUser {
 
     public static void main(String[] args) {
-        System.out.println(AbstractUserMapper.getInstance().map(args[1]));
+        String email = AbstractUserMapper.getInstance().map(args[1]);
+        System.out.println(email);
     }
 
 }